refactor files to be in separate package
