--- src/main/kotlin/org/bibletranslationtools/fetcher/FileType.kt
@@ -1,6 +1,0 @@
-package org.bibletranslationtools.fetcher
-
-enum class FileType {
-    MP3,
-    TR;
-}
