+++ src/main/kotlin/org/bibletranslationtools/fetcher/filesystemaccess/FileType.kt
@@ -1,0 +1,6 @@
+package org.bibletranslationtools.fetcher.filesystemaccess
+
+enum class FileType {
+    MP3,
+    TR;
+}
