+++ src/main/kotlin/org/bibletranslationtools/fetcher/filesystemaccess/Language.kt
@@ -1,4 +1,4 @@
-package org.bibletranslationtools.fetcher
+package org.bibletranslationtools.fetcher.filesystemaccess
 
 data class Language(
     val languageCode: String,
@@ -7,7 +7,7 @@ data class Language(
 ) {
     companion object {
         fun getLanguageByLanguageCode(languageCode: String): Language? {
-            // TODO: implement this using langnames.json
+            // TODO: implement this using language API
             return null
         }
     }
